Replace region preference with room, PR review improvements
