+++ app/src/main/java/com/developersbreach/composeactors/ui/components/RegionEditDialog.kt
@@ -16,6 +16,8 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Close
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
@@ -23,16 +25,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Popup
 import com.developersbreach.composeactors.R
-import java.util.*
+import com.developersbreach.composeactors.utils.getCountryListWithCode
 
 @Composable
-fun RegionEditDialog(popupState: MutableState<Boolean>, onRegionSelect: (String, String) -> Unit) {
+fun RegionEditDialog(
+    popupState: MutableState<Boolean>,
+    onRegionSelect: (String, String) -> Unit
+) {
     if (popupState.value) {
         val regions = getCountryListWithCode()
         Popup(alignment = Alignment.Center, onDismissRequest = { popupState.value = false }) {
@@ -98,24 +102,13 @@ fun Header(popupState: MutableState<Boolean>) {
             style = MaterialTheme.typography.subtitle1
         )
         Image(
-            painter = painterResource(id = R.drawable.ic_close),
+            imageVector = Icons.Rounded.Close,
             contentDescription = stringResource(R.string.close_pop_up_icon),
             modifier = Modifier.clickable {
                 popupState.value = false
             }
         )
     }
-}
-
-private fun getCountryListWithCode(): List<String> {
-    val region = arrayListOf<String>()
-    val isoCountryCodes: Array<String> = Locale.getISOCountries()
-    for (countryCode in isoCountryCodes) {
-        val locale = Locale("", countryCode)
-        val countryName: String = locale.displayCountry
-        region.add("$countryCode,$countryName")
-    }
-    return region
 }
 
 @Composable
